LayoutTests/fast/loader/image-src-multiple-space.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==25');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the image below is displayed.</div>
    <img id="foo" src="resources/image  space.png">
</body>
</html>
